--- src/main/java/de/vfh/workhourstracker/controller/ProjectManagementController.java
@@ -1,19 +1,0 @@
-package de.vfh.workhourstracker.controller;
-
-import de.vfh.workhourstracker.service.ProjectManagementService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/projectManagement")
-public class ProjectManagementController {
-
-    private final ProjectManagementService projectManagementService;
-
-    @Autowired
-    public ProjectManagementController(ProjectManagementService projectManagementService) {
-        this.projectManagementService = projectManagementService;
-    }
-
-}
