--- src/main/java/de/vfh/workhourstracker/controller/ReportingController.java
@@ -1,18 +1,0 @@
-package de.vfh.workhourstracker.controller;
-
-import de.vfh.workhourstracker.service.ReportingService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/reporting")
-public class ReportingController {
-
-    private final ReportingService reportingService;
-
-    @Autowired
-    public ReportingController(ReportingService reportingService) {
-        this.reportingService = reportingService;
-    }
-}
