--- src/main/java/de/vfh/workhourstracker/controller/UserController.java
@@ -1,18 +1,0 @@
-package de.vfh.workhourstracker.controller;
-
-import de.vfh.workhourstracker.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/user")
-public class UserController {
-
-    private final UserService userService;
-
-    @Autowired
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-}
