+++ src/main/java/de/vfh/workhourstracker/reporting/infrastructure/repositories/ReportRepository.java
@@ -1,0 +1,9 @@
+package de.vfh.workhourstracker.reporting.infrastructure.repositories;
+
+import de.vfh.workhourstracker.reporting.domain.model.Report;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReportRepository extends JpaRepository<Report, Long> {
+}
