--- src/main/java/de/vfh/workhourstracker/repository/ProjectRepository.java
@@ -1,9 +1,0 @@
-package de.vfh.workhourstracker.repository;
-
-import de.vfh.workhourstracker.entity.Project;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProjectRepository extends JpaRepository<Project, Long> {
-}
