--- src/main/java/de/vfh/workhourstracker/repository/ReportRepository.java
@@ -1,9 +1,0 @@
-package de.vfh.workhourstracker.repository;
-
-import de.vfh.workhourstracker.entity.Report;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReportRepository extends JpaRepository<Report, Long> {
-}
