--- src/main/java/de/vfh/workhourstracker/repository/TaskRepository.java
@@ -1,9 +1,0 @@
-package de.vfh.workhourstracker.repository;
-
-import de.vfh.workhourstracker.entity.Task;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TaskRepository extends JpaRepository<Task, Long> {
-}
