--- src/main/java/de/vfh/workhourstracker/repository/TimeEntryRepository.java
@@ -1,9 +1,0 @@
-package de.vfh.workhourstracker.repository;
-
-import de.vfh.workhourstracker.entity.TimeEntry;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TimeEntryRepository extends JpaRepository<TimeEntry, Long> {
-}
