--- src/main/java/de/vfh/workhourstracker/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package de.vfh.workhourstracker.repository;
-
-import de.vfh.workhourstracker.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
