+++ src/main/java/de/vfh/workhourstracker/ui/controller/ReportingController.java
@@ -1,0 +1,18 @@
+package de.vfh.workhourstracker.ui.controller;
+
+import de.vfh.workhourstracker.reporting.application.services.ReportingService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/reporting")
+public class ReportingController {
+
+    private final ReportingService reportingService;
+
+    @Autowired
+    public ReportingController(ReportingService reportingService) {
+        this.reportingService = reportingService;
+    }
+}
