+++ src/main/java/de/vfh/workhourstracker/ui/controller/UserController.java
@@ -1,0 +1,18 @@
+package de.vfh.workhourstracker.ui.controller;
+
+import de.vfh.workhourstracker.usermanagement.application.services.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/user")
+public class UserController {
+
+    private final UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+}
