+++ src/test/java/de/vfh/workhourstracker/service/ProjectManagementServiceTest.java
@@ -1,13 +1,12 @@
 package de.vfh.workhourstracker.service;
 
 
+import de.vfh.workhourstracker.projectmanagement.application.services.ProjectManagementService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
