+++ src/test/java/de/vfh/workhourstracker/service/ReportingServiceTest.java
@@ -1,7 +1,6 @@
 package de.vfh.workhourstracker.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import de.vfh.workhourstracker.reporting.application.services.ReportingService;
 
 //@SpringBootTest
 public class ReportingServiceTest {
